requirements.txt
asgiref==3.8.1
channels==4.2.2
channels_redis==4.2.1
Django==5.2.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
msgpack==1.1.1
pillow==11.2.1
PyJWT==2.9.0
pymongo==4.13.2
redis==6.2.0
sqlparse==0.5.3
